requirements/development.txt
-r base.txt

# Development and debugging
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.18.1
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
black==23.12.1
flake8==6.1.0
isort==5.13.2
pre-commit==3.6.0 